cattle-count-wrapper/requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
coloredlogs==15.0.1
Flask==3.0.3
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
h11==0.14.0
humanfriendly==10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.3.0
protobuf==5.27.1
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
sympy==1.12.1
Werkzeug==3.0.3
wsproto==1.2.0
